refactor: déplacer settings vers features/settings/ + supprimer ApiService.ts
