+++ src/features/settings/SettingsPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { SettingsPageView } from '@/views/SettingsPageView';
-import { useSettingsPage } from '@/hooks/useSettingsPage';
+import { SettingsPageView } from './SettingsPageView';
+import { useSettingsPage } from './useSettingsPage';
 
 export default function SettingsPage() {
   const logic = useSettingsPage();
