--- src/services/ApiService.ts
@@ -1,307 +1,0 @@
-// API service to connect frontend with backend database
-import {
-  Player,
-  Game,
-  GameExpansion,
-  GameCharacter,
-  GamePlay,
-  GameFormData,
-  CreatePlayPayload
-} from '@/types';
-
-class ApiService {
-  private readonly baseUrl: string;
-  private readonly TOKEN_KEY = 'auth_token';
-
-  constructor() {
-    this.baseUrl = import.meta.env.VITE_API_URL || 'http://localhost:3001/api';
-  }
-
-  private getToken(): string | null {
-    return localStorage.getItem(this.TOKEN_KEY);
-  }
-
-  private async request<T>(endpoint: string, options: globalThis.RequestInit = {}): Promise<T> {
-    const url = `${this.baseUrl}${endpoint}`;
-    const token = this.getToken();
-
-    const config: globalThis.RequestInit = {
-      method: 'GET',
-      ...options,
-      headers: {
-        'Content-Type': 'application/json',
-        ...(token ? { Authorization: `Bearer ${token}` } : {}),
-        ...(options.headers as Record<string, string> | undefined),
-      },
-    };
-
-    const response = await fetch(url, config);
-
-    if (response.status === 401) {
-      localStorage.removeItem(this.TOKEN_KEY);
-      throw new Error('UNAUTHORIZED');
-    }
-
-    if (!response.ok) {
-      const errorData = await response.json().catch(() => ({ error: 'Unknown error' }));
-      throw new Error(errorData.error || `HTTP ${response.status}: ${response.statusText}`);
-    }
-
-    // Handle 204 No Content responses
-    if (response.status === 204) {
-      return {} as T;
-    }
-
-    return await response.json();
-  }
-
-  // Auth
-  async login(password: string): Promise<void> {
-    const response = await fetch(`${this.baseUrl}/auth/login`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ password }),
-    });
-    if (!response.ok) throw new Error('Invalid password');
-    const data = await response.json() as { token: string };
-    localStorage.setItem(this.TOKEN_KEY, data.token);
-  }
-
-  logout(): void {
-    localStorage.removeItem(this.TOKEN_KEY);
-  }
-
-  isAuthenticated(): boolean {
-    return this.getToken() !== null;
-  }
-
-  // Player operations
-  async getAllPlayers() {
-    return this.request<Player[]>('/players');
-  }
-
-  async getPlayerById(playerId: number) {
-    return this.request<Player>(`/players/${playerId}`);
-  }
-
-  async createPlayer(playerData: { player_name: string; avatar?: string; favorite_game?: string }) {
-    return this.request<Player>('/players', {
-      method: 'POST',
-      body: JSON.stringify(playerData),
-    });
-  }
-
-  async updatePlayer(playerId: number, playerData: Partial<Player>) {
-    return this.request<Player>(`/players/${playerId}`, {
-      method: 'PUT',
-      body: JSON.stringify(playerData),
-    });
-  }
-
-  async deletePlayer(playerId: number) {
-    return this.request<void>(`/players/${playerId}`, {
-      method: 'DELETE',
-    });
-  }
-
-  // Game operations
-  async getAllGames() {
-    return this.request<Game[]>('/games');
-  }
-
-  async getGameById(gameId: number) {
-    return this.request<Game>(`/games/${gameId}`);
-  }
-
-  async createGame(gameData: GameFormData) {
-    const sanitized = {
-      ...gameData,
-      image: gameData.image || undefined,
-      thumbnail: gameData.thumbnail || undefined,
-    };
-    return this.request<Game>('/games', {
-      method: 'POST',
-      body: JSON.stringify(sanitized),
-    });
-  }
-
-  async updateGame(gameId: number, gameData: Partial<GameFormData>) {
-    const sanitized = {
-      ...gameData,
-      image: gameData.image || undefined,
-      thumbnail: gameData.thumbnail || undefined,
-    };
-    return this.request<Game>(`/games/${gameId}`, {
-      method: 'PUT',
-      body: JSON.stringify(sanitized),
-    });
-  }
-
-  async deleteGame(gameId: number) {
-    return this.request<void>(`/games/${gameId}`, {
-      method: 'DELETE',
-    });
-  }
-
-  // Game Expansion operations
-  async getAllExpansions(gameId?: number) {
-    const query = gameId ? `?game_id=${gameId}` : '';
-    return this.request<GameExpansion[]>(`/expansions${query}`);
-  }
-
-  async getExpansionById(expansionId: number) {
-    return this.request<GameExpansion>(`/expansions/${expansionId}`);
-  }
-
-  async createExpansion(expansionData: Omit<GameExpansion, 'expansion_id'>) {
-    return this.request<GameExpansion>('/expansions', {
-      method: 'POST',
-      body: JSON.stringify(expansionData),
-    });
-  }
-
-  async updateExpansion(expansionId: number, expansionData: Omit<GameExpansion, 'expansion_id' | 'game_id'>) {
-    return this.request<GameExpansion>(`/expansions/${expansionId}`, {
-      method: 'PUT',
-      body: JSON.stringify(expansionData),
-    });
-  }
-
-  async deleteExpansion(expansionId: number) {
-    return this.request<void>(`/expansions/${expansionId}`, {
-      method: 'DELETE',
-    });
-  }
-
-  // Game Character operations
-  async getAllCharacters(gameId?: number) {
-    const query = gameId ? `?game_id=${gameId}` : '';
-    return this.request<GameCharacter[]>(`/characters${query}`);
-  }
-
-  async getCharacterById(characterId: number) {
-    return this.request<GameCharacter>(`/characters/${characterId}`);
-  }
-
-  async createCharacter(characterData: Omit<GameCharacter, 'character_id'>) {
-    return this.request<GameCharacter>('/characters', {
-      method: 'POST',
-      body: JSON.stringify(characterData),
-    });
-  }
-
-  async updateCharacter(characterId: number, characterData: Omit<GameCharacter, 'character_id' | 'game_id'>) {
-    return this.request<GameCharacter>(`/characters/${characterId}`, {
-      method: 'PUT',
-      body: JSON.stringify(characterData),
-    });
-  }
-
-  async deleteCharacter(characterId: number) {
-    return this.request<void>(`/characters/${characterId}`, {
-      method: 'DELETE',
-    });
-  }
-
-  async createPlay(playData: CreatePlayPayload) {
-    return this.request<GamePlay>('/plays', {
-      method: 'POST',
-      body: JSON.stringify(playData),
-    });
-  }
-
-  // Play operations
-  async getAllPlays(gameId?: number) {
-    const query = gameId ? `?game_id=${gameId}` : '';
-    return this.request<GamePlay[]>(`/plays${query}`);
-  }
-
-  async getPlaysByGame(gameId: number) {
-    return this.request<GamePlay[]>(`/plays/game/${gameId}`);
-  }
-
-  async getPlaysByPlayer(playerId: number) {
-    return this.request<GamePlay[]>(`/plays/player/${playerId}`);
-  }
-
-  async updatePlay(playId: number, playData: Partial<GamePlay>) {
-    return this.request<GamePlay>(`/plays/${playId}`, {
-      method: 'PUT',
-      body: JSON.stringify(playData),
-    });
-  }
-
-  async deletePlay(playId: number) {
-    return this.request<void>(`/plays/${playId}`, {
-      method: 'DELETE',
-    });
-  }
-
-  // Statistics
-  async getPlayerStats() {
-    return this.request<unknown>('/stats/players');
-  }
-
-  async getGameStats() {
-    return this.request<unknown>('/stats/games');
-  }
-
-  // Data export / import / reset (admin only)
-  async exportData(): Promise<Blob> {
-    const url = `${this.baseUrl}/v1/data/export`;
-    const token = this.getToken();
-    const res = await fetch(url, {
-      headers: token ? { Authorization: `Bearer ${token}` } : {},
-    });
-    if (res.status === 401) {
-      localStorage.removeItem(this.TOKEN_KEY);
-      throw new Error('UNAUTHORIZED');
-    }
-    if (!res.ok) throw new Error(`Export failed: ${res.status}`);
-    return res.blob();
-  }
-
-  async importData(file: File): Promise<{ ok: boolean }> {
-    const text = await file.text();
-    return this.request<{ ok: boolean }>('/v1/data/import', {
-      method: 'POST',
-      body: text,
-    });
-  }
-
-  async resetData(): Promise<{ ok: boolean }> {
-    return this.request<{ ok: boolean }>('/v1/data/reset', { method: 'POST' });
-  }
-
-  // Health check
-  async healthCheck() {
-    return this.request<{ status: string; timestamp: string }>('/health');
-  }
-
-  // BGG Catalog status (count + last import date)
-  async getBggCatalogStatus(): Promise<{ count: number; bgg_catalog_imported_at: string | null }> {
-    return this.request<{ count: number; bgg_catalog_imported_at: string | null }>('/v1/bgg/import-status');
-  }
-
-  async importBggCatalog(file: File): Promise<{ count: number }> {
-    const text = await file.text();
-    const url = `${this.baseUrl}/v1/bgg/import-catalog`;
-    const response = await fetch(url, {
-      method: 'POST',
-      credentials: 'include',
-      headers: { 'Content-Type': 'text/plain' },
-      body: text,
-    });
-    if (response.status === 401) {
-      localStorage.removeItem(this.TOKEN_KEY);
-      throw new Error('UNAUTHORIZED');
-    }
-    if (!response.ok) {
-      const errorData = await response.json().catch(() => ({ error: 'Unknown error' }));
-      throw new Error(errorData.error || `HTTP ${response.status}`);
-    }
-    return response.json();
-  }
-}
-
-export default new ApiService();
